Separate camp DB/separate camp utils. Types corrections.
