+++ src/engine/core/schemes/animpoint/index.ts
@@ -1,3 +1,2 @@
-export * from "@/engine/core/schemes/animpoint/ISchemeAnimpointState";
 export * from "@/engine/core/schemes/animpoint/SchemeAnimpoint";
 export * from "@/engine/core/schemes/animpoint/types";
