+++ src/engine/core/schemes/animpoint/types.ts
@@ -1,5 +1,22 @@
 import { EStalkerState } from "@/engine/core/objects/animation";
-import { ClientObject, LuaArray, TName } from "@/engine/lib/types";
+import { IBaseSchemeState } from "@/engine/core/schemes";
+import { AnimpointManager } from "@/engine/core/schemes/animpoint/AnimpointManager";
+import { ClientObject, LuaArray, Optional, TDistance, TName } from "@/engine/lib/types";
+
+/**
+ * State of animpoint scheme.
+ */
+export interface ISchemeAnimpointState extends IBaseSchemeState {
+  animpoint: AnimpointManager;
+  actionNameBase: Optional<TName>;
+  coverName: TName;
+  useCamp: boolean;
+  reachDistance: TDistance; // Already squared.
+  reachMovement: EStalkerState;
+  description: Optional<EStalkerState>;
+  availableAnimations: Optional<LuaArray<EStalkerState>>;
+  approvedActions: LuaArray<IAnimpointActionDescriptor>;
+}
 
 /**
  * Descriptor of animpoint when object is captured in smart cover and deciding which specific animation to run.
@@ -7,13 +24,4 @@ import { ClientObject, LuaArray, TName } from "@/engine/lib/types";
 export interface IAnimpointActionDescriptor {
   name: EStalkerState;
   predicate: (this: void, object: ClientObject, isInCamp?: boolean) => boolean;
-}
-
-/**
- * Descriptor of animations when objects are participating in camp scenario.
- * One is telling, others are listening.
- */
-export interface IStoryAnimationDescriptor {
-  director: LuaArray<TName>;
-  listener: LuaArray<TName>;
 }
