+++ src/engine/core/schemes/walker/actions/ActionWalkerActivity.ts
@@ -1,6 +1,7 @@
 import { action_base, LuabindClass } from "xray16";
 
 import { registry, setStalkerState } from "@/engine/core/database";
+import { getCampForPosition } from "@/engine/core/database/camp";
 import { GlobalSoundManager } from "@/engine/core/managers/sounds/GlobalSoundManager";
 import { EStalkerState } from "@/engine/core/objects/animation";
 import { animpoint_predicates } from "@/engine/core/objects/animation/predicates/animpoint_predicates";
@@ -89,7 +90,7 @@ export class ActionWalkerActivity extends action_base implements ISchemeEventHan
 
     this.moveManager.update();
 
-    const camp = CampManager.getInstanceForPosition(this.object.position());
+    const camp: Optional<CampManager> = getCampForPosition(this.object.position());
 
     if (camp !== null && this.state.use_camp === true) {
       this.campStoryManager = camp;
